+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Versions
 
+## 1.6.2
+
+- Fix 'invalid asset' bug
+- AutoSwap: disallow spending LBTC gained in the same channel
+- AutoSwap: only consider channels with routed out > routed in
+
 ## 1.6.1
 
 - AutoFee: add Update All to set paramereter(s) to all custom rules